aibro-social-network/requirements.txt
django  # For Django itself
djangorestframework  # For REST Framework
django-cors-headers  # For CORS
djangorestframework-simplejwt  # For JWT authentication
python-dotenv  # For loading .env files
Pillow  # For image handling (ImageField)
drf-nested-routers
django-allauth
dj-rest-auth
django-cors-headers
django-templated-mail
requests
cryptography
google-auth
channels
channels-redis
pyfcm
exponent_server_sdk